+++ lua/plugins/gitsigns.lua
@@ -1,0 +1,9 @@
+-- lua/plugins/gitsigns.lua
+
+return {
+    'lewis6991/gitsigns.nvim',
+    event  = { 'BufReadPre', 'BufNewFile' },
+    config = function()
+        require('config.gitsigns')
+    end,
+}